+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0]
+
+### Added
+
+- new themes of BaseExpander
+- new theme of BaseIconButton
+
+### Changed
+
+- BaseExpander docs about slot 'icon'
+- BaseExpander styles of arrow icon
+
+### Fixed
+
+- visibility of text and icon in disabled BaseButton in dark theme
+
 ## [0.7.0]
 
 ### Added